requirements.txt
apimatic-core==0.2.18
apimatic-core-interfaces==0.1.6
apimatic-requests-client-adapter==0.1.7
asgiref==3.8.1
attrs==25.3.0
boto3==1.37.2
botocore==1.37.2
CacheControl==0.12.14
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dataclasses-json==0.6.7
deprecation==2.1.0
Django==5.1.6
django-allauth==65.7.0
django-storages==1.14.5
h11==0.14.0
idna==3.10
jmespath==1.0.1
jsonpath-python==1.0.6
jsonpickle==3.3.0
jsonpointer==2.4
jwt==1.3.1
marshmallow==3.26.1
msgpack==1.1.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.3
selenium==4.30.0
setuptools==75.8.0
shippo==3.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
squareup==40.0.0.220250123
trio==0.29.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
